Avoid exposing primary index methods on models including WithoutPrimaryIndex
