+++ lib/identity_cache/query_api.rb
@@ -8,97 +8,9 @@ module IdentityCache
     end
 
     module ClassMethods
-      # Similar to ActiveRecord::Base#exists? will return true if the id can be
-      # found in the cache or in the DB.
-      def exists_with_identity_cache?(id)
-        unless primary_cache_index_enabled
-          raise NotImplementedError, "exists_with_identity_cache? needs the primary index enabled"
-        end
-        !!fetch_by_id(id)
-      end
-
-      # Default fetcher added to the model on inclusion, it behaves like
-      # ActiveRecord::Base.where(id: id).first
-      def fetch_by_id(id, includes: nil)
-        ensure_base_model
-        raise_if_scoped
-        raise NotImplementedError, "fetching needs the primary index enabled" unless primary_cache_index_enabled
-        id = type_for_attribute(primary_key).cast(id)
-        return unless id
-        record = if should_use_cache?
-          object = nil
-          coder  = IdentityCache.fetch(rails_cache_key(id)) do
-            Encoder.encode(object = resolve_cache_miss(id))
-          end
-          object ||= Encoder.decode(coder, self)
-          if object && object.id != id
-            IdentityCache.logger.error(
-              <<~MSG.squish
-                [IDC id mismatch] fetch_by_id_requested=#{id}
-                fetch_by_id_got=#{object.id}
-                for #{object.inspect[(0..100)]}
-              MSG
-            )
-          end
-          object
-        else
-          resolve_cache_miss(id)
-        end
-        prefetch_associations(includes, [record]) if record && includes
-        record
-      end
-
-      # Default fetcher added to the model on inclusion, it behaves like
-      # ActiveRecord::Base.find, will raise ActiveRecord::RecordNotFound exception
-      # if id is not in the cache or the db.
-      def fetch(id, includes: nil)
-        fetch_by_id(id, includes: includes) or raise(
-          ActiveRecord::RecordNotFound, "Couldn't find #{name} with ID=#{id}"
-        )
-      end
-
-      # Default fetcher added to the model on inclusion, if behaves like
-      # ActiveRecord::Base.find_all_by_id
-      def fetch_multi(*ids, includes: nil)
-        ensure_base_model
-        raise_if_scoped
-        raise NotImplementedError, "fetching needs the primary index enabled" unless primary_cache_index_enabled
-        ids.flatten!(1)
-        id_type = type_for_attribute(primary_key)
-        ids.map! { |id| id_type.cast(id) }.compact!
-        records = if should_use_cache?
-          cache_keys = ids.map { |id| rails_cache_key(id) }
-          key_to_id_map = Hash[ cache_keys.zip(ids) ]
-          key_to_record_map = {}
-
-          coders_by_key = IdentityCache.fetch_multi(cache_keys) do |unresolved_keys|
-            ids = unresolved_keys.map { |key| key_to_id_map[key] }
-            records = find_batch(ids)
-            key_to_record_map = records.compact.index_by { |record| rails_cache_key(record.id) }
-            records.map { |record| Encoder.encode(record) }
-          end
-
-          cache_keys.map do |key|
-            key_to_record_map[key] || Encoder.decode(coders_by_key[key], self)
-          end
-        else
-          find_batch(ids)
-        end
-        records.compact!
-        prefetch_associations(includes, records) if includes
-        records
-      end
-
       # Prefetches cached associations on a collection of records
       def prefetch_associations(includes, records)
         Cached::Prefetcher.prefetch(self, includes, records)
-      end
-
-      # Invalidates the primary cache index for the associated record. Will not invalidate cached attributes.
-      def expire_primary_key_cache_index(id)
-        return unless primary_cache_index_enabled
-        id = type_for_attribute(primary_key).cast(id)
-        IdentityCache.cache.delete(rails_cache_key(id))
       end
 
       # @api private
@@ -124,12 +36,6 @@ module IdentityCache
             scoped with a join isn't supported
           MSG
         end
-      end
-
-      def resolve_cache_miss(id)
-        record = includes(cache_fetch_includes).where(primary_key => id).take
-        setup_embedded_associations_on_miss([record]) if record
-        record
       end
 
       def preload_id_embedded_association(records, cached_association)
@@ -245,22 +151,10 @@ module IdentityCache
 
         associations_for_identity_cache.compact
       end
-
-      def find_batch(ids)
-        return [] if ids.empty?
-
-        @id_column ||= columns.detect { |c| c.name == primary_key }
-        ids = ids.map { |id| connection.type_cast(id, @id_column) }
-        records = where(primary_key => ids).includes(cache_fetch_includes).to_a
-        setup_embedded_associations_on_miss(records)
-        records_by_id = records.index_by(&:id)
-        ids.map { |id| records_by_id[id] }
-      end
     end
 
     # Invalidate the cache data associated with the record.
     def expire_cache
-      expire_primary_index
       expire_attribute_indexes
       true
     end
@@ -322,10 +216,6 @@ module IdentityCache
           inverse_cached_association.records_variable_name, self
         )
       end
-    end
-
-    def expire_primary_index # :nodoc:
-      self.class.expire_primary_key_cache_index(id)
     end
 
     def expire_attribute_indexes # :nodoc:
