+++ lib/identity_cache/without_primary_index.rb
@@ -3,9 +3,29 @@ module IdentityCache
   module WithoutPrimaryIndex
     extend ActiveSupport::Concern
 
-    included do |base|
-      base.send(:include, IdentityCache)
-      base.primary_cache_index_enabled = false
+    include ArTransactionChanges
+    include IdentityCache::BelongsToCaching
+    include IdentityCache::CacheKeyGeneration
+    include IdentityCache::ConfigurationDSL
+    include IdentityCache::QueryAPI
+    include IdentityCache::CacheInvalidation
+    include IdentityCache::ShouldUseCache
+    include ParentModelExpiration
+
+    def self.append_features(base) #:nodoc:
+      raise AlreadyIncludedError if base.include?(WithoutPrimaryIndex)
+      super
+    end
+
+    included do
+      class_attribute(:cached_model)
+      self.cached_model = self
+    end
+
+    module ClassMethods
+      def primary_cache_index_enabled
+        false
+      end
     end
   end
 end
